requirements.txt
langchain-openai>=0.1.0
python-dotenv>=1.0.0
elasticsearch>=8.0.0
numpy>=1.24.0
openai>=1.0.0
